requirements.txt
mcp>=1.21.2
pymssql>=2.2.7
python-dotenv>=1.2.1
mysql-connector-python
